cast bit_swap result back to input type

+++ src/mod/src/controls/z_jump.cpp
@@ -22,7 +22,7 @@ extern "C" void hook_PADRead(PADStatus *status)
 
 		// Swap X/Y bit with Z bit
 		if (const auto xy_bit = controller_configs[i].z_jump_bit; xy_bit != 0)
-			status[i].buttons = (u16)bit_swap(status[i].buttons, xy_bit, ButtonBit_Z);
+			status[i].buttons = bit_swap(status[i].buttons, xy_bit, ButtonBit_Z);
 	}
 }
